Update privacy-policy.md
Added that we process company information for plans for processional use, that our customer service team try to become better by evaluating their dialogue with customers, that we process usage data for security and fault/error detection purposes, and that we may share information with co-marketing content provider if we at the same time make it clear to the user that we will be sharing the content with said provider. Finally, the text concerning product and marketing communication has been future-proofed so we in the future can provide more specific information to customers, and we have clarified that we do not ask for consent if it is not necessary.